Model working! Not predicting super good but working!
Model effnet_v2B1_lr0.0005_size128_y2016-2018-2020-2022-2024_stack1-4_Pooling was trained under this specification 👍

+++ src/geo_utils.py
@@ -440,7 +440,7 @@ def random_tiled_image_from_census_tract(
     return composition, point, boundaries, total_boundaries
 
 
-def process_image(img, resizing_size, moveaxis=True):
+def process_image(img, resizing_size, moveaxis=True, dtype=np.uint8):
     if moveaxis:
         img = np.moveaxis(
             img, 0, 2
@@ -454,6 +454,7 @@ def process_image(img, resizing_size, moveaxis=True):
             img, dsize=(resizing_size, resizing_size), interpolation=cv2.INTER_CUBIC
         )
 
+    img = img.astype(dtype)
     return img
